SQLContextSingleton.getInstance INSIDE rdd.collect [FAIL]

14:02:36
[Stage 52:> (0 + 2) / 8] [Stage 52:=======> (1 + 2) / 8] [Stage
52:==============> (2 + 2) / 8] [Stage 52:=====================> (3 + 2

14:02:36
(1) RDD transformations and actions are NOT invoked by the driver, but
inside of other transformations; for example, rdd1.map(x =>
rdd2.values.count() * x) is invalid because the values transformation
and count action cannot be performed inside of the rdd1.map
transformation. For more information, see SPARK-5063.

14:02:36
(2) When a Spark Streaming job recovers from checkpoint, this exception
will be hit if a reference to an RDD not defined by the streaming job
is used in DStream operations. For more information, See SPARK-13758.